+++ R/direct_messages.R
@@ -6,6 +6,7 @@
 #' about the sender and recipient.
 #'
 #' @inheritParams TWIT_paginate_cursor
+#' @inheritParams stream
 #' @param next_cursor `r lifecycle::badge("deprecated")` Use `cursor` instead.
 #' @return A list with one element for each page of results.
 #' @examples