+++ R/followers.R
@@ -5,6 +5,7 @@
 #'
 #' @inheritParams TWIT_paginate_cursor
 #' @inheritParams get_timeline
+#' @inheritParams stream
 #' @param page `r lifecycle::badge("deprecated")` Please use `cursor` instead.
 #' @references <https://developer.twitter.com/en/docs/twitter-api/v1/accounts-and-users/follow-search-get-users/api-reference/get-followers-ids>
 #' @examples