+++ R/lists_subscribers.R
@@ -1,8 +1,9 @@
 #' Get subscribers of a specified list.
 #'
 #' @inheritParams TWIT_paginate_cursor
+#' @inheritParams stream
 #' @param list_id required The numerical id of the list.
-#' @param slug,owner_user The list name (slug) and owner. 
+#' @param slug,owner_user The list name (slug) and owner.
 #' @examples
 #' if (auth_has_default()) {
 #'     ## get subscribers of rladies list
@@ -29,9 +30,9 @@ lists_subscribers <- function(list_id = NULL,
     owner_user = owner_user
   )
 
-  r <- TWIT_paginate_cursor(token, "/1.1/lists/subscribers", params, 
+  r <- TWIT_paginate_cursor(token, "/1.1/lists/subscribers", params,
     n = n,
-    cursor = cursor, 
+    cursor = cursor,
     retryonratelimit = retryonratelimit,
     verbose = verbose,
     page_size = if (n >= 5000) 5000 else n,
@@ -49,6 +50,6 @@ parse_lists_users <- function(x) {
   dfs <- lapply(users, wrangle_into_clean_data, type = "user")
   dfs <- lapply(dfs, tibble::as_tibble)
   df <- do.call(rbind, dfs)
-  
+
   copy_cursor(df, x)
 }